refactor(arch3-core): move the rewards queries to a facade
